reproducible test case

+++ hello-world.cc
@@ -2,8 +2,8 @@
 #include <clasp/clasp.h>
 
 namespace hw {
-
-void hello_world(double x, double y) { fmt::print("Hello World! Did you know that {} + {} = {}?\n", x, y, x + y); }
+// causes LOAD-FOREIGN-LIBRARY-ERROR
+void hello_world(double x, uint64_t y) { fmt::print("Hello World! Did you know that {} + {} = {}?\n", x, y, x + y);}
 
 extern "C" void startup_clasp_extension() {
   using namespace clbind;
